Moved code back in one file. INSTALL
This will easy installation procedure / packaging
